Bring code in line with pachyderm standards.
